KTO-1423: Näytetään koulutustyyppi  lyhyempänä etusivun listoissa
