+++ src/main/app/src/components/KoulutustyyppiSelect/index.tsx
@@ -9,53 +9,13 @@ import SegmentTabs from '#/src/components/SegmentTabs';
 import { Box, Radio, RadioGroup } from '#/src/components/virkailija';
 import {
   KOULUTUSTYYPIT,
-  KOULUTUSTYYPPI,
-  TUTKINTOON_JOHTAMATTOMAT_KORKEAKOULU_KOULUTUSTYYPIT,
-  TUTKINTOON_JOHTAVAT_KORKEAKOULU_KOULUTUSTYYPIT,
   TUTKINTOON_JOHTAVAT_KOULUTUSTYYPIT,
-  TUTKINTOON_JOHTAMATTOMAT_AMMATILLISET_KOULUTUSTYYPIT,
-  getKoulutustyyppiTranslationKey,
+  TUTKINTOON_JOHTAVA_KOULUTUSTYYPPIHIERARKIA,
+  TUTKINTOON_JOHTAMATON_KOULUTUSTYYPPIHIERARKIA,
 } from '#/src/constants';
 import { spacing, getThemeProp } from '#/src/theme';
-import { getTestIdProps } from '#/src/utils';
+import { getTestIdProps, getKoulutustyyppiTranslationKey } from '#/src/utils';
 import iterateTree, { Order } from '#/src/utils/iterateTree';
-
-export const TUTKINTOON_JOHTAVA_KOULUTUSTYYPPIHIERARKIA = [
-  { value: KOULUTUSTYYPPI.AMMATILLINEN_KOULUTUS },
-  {
-    value: 'korkeakoulutus',
-    children: TUTKINTOON_JOHTAVAT_KORKEAKOULU_KOULUTUSTYYPIT.map(kt => ({
-      value: kt,
-    })),
-  },
-  { value: KOULUTUSTYYPPI.LUKIOKOULUTUS },
-];
-
-export const TUTKINTOON_JOHTAMATON_KOULUTUSTYYPPIHIERARKIA = [
-  {
-    value: 'ammatillinen',
-    children: TUTKINTOON_JOHTAMATTOMAT_AMMATILLISET_KOULUTUSTYYPIT.map(kt => ({
-      value: kt,
-    })),
-  },
-  {
-    value: 'korkeakoulutus',
-    children: TUTKINTOON_JOHTAMATTOMAT_KORKEAKOULU_KOULUTUSTYYPIT.map(kt => ({
-      value: kt,
-    })),
-  },
-  {
-    value: 'vapaa-sivistystyo',
-    children: [
-      {
-        value: KOULUTUSTYYPPI.VAPAA_SIVISTYSTYO_OPISTOVUOSI,
-      },
-      { value: KOULUTUSTYYPPI.VAPAA_SIVISTYSTYO_MUU },
-    ],
-  },
-  { value: KOULUTUSTYYPPI.TUVA },
-  { value: KOULUTUSTYYPPI.PERUSOPETUKSEN_LISAOPETUS },
-];
 
 const SecondLevelContainer = styled(Box).attrs({ flexGrow: 0 })`
   margin-left: ${spacing(4)};
